+++ selendroid-standalone/src/test/java/io/selendroid/android/impl/DefaultAndroidEmulatorTests.java
@@ -18,6 +18,8 @@ import io.selendroid.android.AndroidEmulator;
 import java.io.File;
 import java.util.List;
 import java.util.Locale;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.Assert;