+++ selendroid-standalone/src/test/java/io/selendroid/server/model/DeviceStoreTest.java
@@ -14,6 +14,7 @@
 package io.selendroid.server.model;
 
 import static io.selendroid.server.model.DeviceStoreFixture.anDevice;
+import static io.selendroid.server.model.DeviceStoreFixture.anDeviceManager;
 import static io.selendroid.server.model.DeviceStoreFixture.anEmulator;
 import static io.selendroid.server.model.DeviceStoreFixture.withDefaultCapabilities;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -37,6 +38,7 @@ import io.selendroid.exceptions.DeviceStoreException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -52,7 +54,7 @@ public class DeviceStoreTest {
   public void shouldAddStartedEmulator() throws Exception {
     AndroidEmulator emulator = anEmulator("de", DeviceTargetPlatform.ANDROID10, false);
 
-    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     deviceStore.addEmulators(Collections.singletonList(emulator));
     Assert.assertEquals(deviceStore.getDevicesInUse().size(), 0);
     Assert.assertEquals(deviceStore.getDevicesList().size(), 1);
@@ -63,8 +65,8 @@ public class DeviceStoreTest {
   }
 
   @Test
-  public void shouldIncrementEmulatorPortsByTwo() {
-    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT);
+  public void shouldIncrementEmulatorPortsByTwo() throws Exception {
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     Assert.assertEquals(5560, deviceStore.nextEmulatorPort().intValue());
     Assert.assertEquals(5562, deviceStore.nextEmulatorPort().intValue());
     Assert.assertEquals(5564, deviceStore.nextEmulatorPort().intValue());
@@ -77,7 +79,7 @@ public class DeviceStoreTest {
 
     EmulatorPortFinder finder = mock(EmulatorPortFinder.class);
     when(finder.next()).thenReturn(5554);
-    DeviceStore deviceStore = new DeviceStore(finder, false);
+    DeviceStore deviceStore = new DeviceStore(finder, false, anDeviceManager());
     deviceStore.addEmulators(Arrays.asList(new AndroidEmulator[] {deEmulator}));
     Assert.assertEquals(deviceStore.getDevicesInUse().size(), 0);
     AndroidDevice foundDevice = deviceStore.findAndroidDevice(withDefaultCapabilities());
@@ -97,7 +99,7 @@ public class DeviceStoreTest {
     AndroidEmulator enEmulator10 = anEmulator("en", DeviceTargetPlatform.ANDROID10, false);
     AndroidEmulator deEmulator16 = anEmulator("de", DeviceTargetPlatform.ANDROID16, false);
 
-    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     deviceStore.addEmulators(Arrays.asList(new AndroidEmulator[] {deEmulator10, enEmulator10,
         deEmulator16}));
     Assert.assertEquals(deviceStore.getDevicesInUse().size(), 0);
@@ -114,37 +116,46 @@ public class DeviceStoreTest {
   }
 
   @Test
-  public void shouldRegisterNotStartedEmulatorAndSkipOthers() throws Exception {
+  public void shouldRegisterStartedAndStoppedEmulators() throws Exception {
     AndroidEmulator deEmulator10 = anEmulator("de", DeviceTargetPlatform.ANDROID10, false);
     AndroidEmulator enEmulator10 = anEmulator("en", DeviceTargetPlatform.ANDROID10, true);
     AndroidEmulator deEmulator16 = anEmulator("de", DeviceTargetPlatform.ANDROID16, true);
 
-    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     deviceStore.addEmulators(Arrays.asList(new AndroidEmulator[] {deEmulator10, enEmulator10,
         deEmulator16}));
     Assert.assertEquals(deviceStore.getDevicesInUse().size(), 0);
-    Assert.assertEquals(deviceStore.getDevicesList().size(), 1);
 
-    Assert.assertTrue(deviceStore.getDevicesList().containsKey(DeviceTargetPlatform.ANDROID10));
-    Assert.assertTrue(deviceStore.getDevicesList().get(DeviceTargetPlatform.ANDROID10)
-        .contains(deEmulator10));
+    Assert.assertTrue("Should have an entry for target version 10.", deviceStore.getDevicesList()
+        .containsKey(DeviceTargetPlatform.ANDROID10));
+    List<AndroidDevice> devicesApi10 =
+        deviceStore.getDevicesList().get(DeviceTargetPlatform.ANDROID10);
+    Assert.assertTrue("Entry of target version 10 should contain de emulator.",
+        devicesApi10.contains(deEmulator10));
+    Assert.assertTrue("Entry of target version 10 should contain en emulator.",
+        devicesApi10.contains(enEmulator10));
+
+    Assert.assertTrue("Should have an entry for target version 16.", deviceStore.getDevicesList()
+        .containsKey(DeviceTargetPlatform.ANDROID16));
+    Assert.assertTrue("Entry of target version 16 should contain the emulator.", deviceStore
+        .getDevicesList().get(DeviceTargetPlatform.ANDROID16).contains(deEmulator16));
   }
 
   @Test
-  public void shouldInitStoreAndIgnoreRunningEmulators() throws Exception {
+  public void shouldNotIgnoreRunningEmulators() throws Exception {
     AndroidEmulator enEmulator10 = anEmulator("en", DeviceTargetPlatform.ANDROID10, true);
     AndroidEmulator deEmulator16 = anEmulator("de", DeviceTargetPlatform.ANDROID16, true);
 
-    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     deviceStore.addEmulators(Arrays.asList(new AndroidEmulator[] {enEmulator10, deEmulator16}));
 
     // Nothing has been added.
-    Assert.assertEquals(deviceStore.getDevicesList().size(), 0);
+    Assert.assertEquals(deviceStore.getDevicesList().size(), 2);
   }
 
   @Test
   public void storeShouldDoNothingIfInitializedWithEmptyList() throws AndroidDeviceException {
-    DeviceStore store = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore store = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
 
     store.addEmulators(new ArrayList<AndroidEmulator>());
     Assert.assertEquals(store.getDevicesList().size(), 0);
@@ -152,7 +163,7 @@ public class DeviceStoreTest {
 
   @Test
   public void storeShouldDoNothingIfInitializedWithNull() throws AndroidDeviceException {
-    DeviceStore store = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore store = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
 
     store.addEmulators(null);
     Assert.assertEquals(store.getDevicesList().size(), 0);
@@ -161,10 +172,10 @@ public class DeviceStoreTest {
 
 
   @Test
-  public void storeShouldBeAbleToFindDeviceForCapabilities() throws Exception {
+  public void shouldFindSwitchedOffEmulator() throws Exception {
     // prepare device store
     DefaultAndroidEmulator deEmulator16 = anEmulator("de", DeviceTargetPlatform.ANDROID16, false);
-    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     deviceStore.addEmulators(Arrays.asList(new AndroidEmulator[] {deEmulator16}));
 
     // find by Capabilities
@@ -179,7 +190,7 @@ public class DeviceStoreTest {
   public void shouldThrowAnExceptionIfTargetPlatformIsMissingInCapabilities() throws Exception {
     // prepare device store
     DefaultAndroidEmulator deEmulator16 = anEmulator("de", DeviceTargetPlatform.ANDROID16, false);
-    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     deviceStore.addEmulators(Arrays.asList(new AndroidEmulator[] {deEmulator16}));
 
     SelendroidCapabilities capa = new SelendroidCapabilities();
@@ -199,7 +210,7 @@ public class DeviceStoreTest {
     // prepare device store
     DefaultAndroidEmulator deEmulator10 = anEmulator("de", DeviceTargetPlatform.ANDROID10, false);
     AndroidEmulator enEmulator10 = anEmulator("en", DeviceTargetPlatform.ANDROID10, false);
-    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     deviceStore.addEmulators(Arrays.asList(new AndroidEmulator[] {deEmulator10, enEmulator10}));
 
     // find by Capabilities
@@ -220,7 +231,7 @@ public class DeviceStoreTest {
     // prepare device store
     DefaultAndroidEmulator deEmulator10 = anEmulator("de", DeviceTargetPlatform.ANDROID16, false);
     AndroidEmulator enEmulator10 = anEmulator("en", DeviceTargetPlatform.ANDROID10, false);
-    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     deviceStore.addEmulators(Arrays.asList(new AndroidEmulator[] {deEmulator10, enEmulator10}));
 
     // find by Capabilities
@@ -245,7 +256,7 @@ public class DeviceStoreTest {
     when(device.getTargetPlatform()).thenReturn(DeviceTargetPlatform.ANDROID16);
     when(device.isDeviceReady()).thenReturn(Boolean.TRUE);
 
-    DeviceStore store = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore store = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     store.addDevice(device);
     assertThat(store.getDevicesList().values(), hasSize(1));
     assertThat(store.getDevicesInUse(), hasSize(0));
@@ -258,7 +269,7 @@ public class DeviceStoreTest {
     when(device.getTargetPlatform()).thenReturn(DeviceTargetPlatform.ANDROID16);
     when(device.isDeviceReady()).thenReturn(Boolean.FALSE);
 
-    DeviceStore store = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore store = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     store.addDevice(device);
     assertThat(store.getDevicesList().values(), hasSize(0));
     assertThat(store.getDevicesInUse(), hasSize(0));
@@ -272,7 +283,7 @@ public class DeviceStoreTest {
     when(device.getScreenSize()).thenReturn("320x480");
     when(device.screenSizeMatches("320x480")).thenReturn(Boolean.TRUE);
 
-    DeviceStore store = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore store = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     store.addDevice(device);
     assertThat(store.getDevicesList().values(), hasSize(1));
     assertThat(store.getDevicesInUse(), hasSize(0));
@@ -291,7 +302,7 @@ public class DeviceStoreTest {
     when(device.getScreenSize()).thenReturn("320x500");
     when(device.screenSizeMatches("320x500")).thenReturn(Boolean.FALSE);
 
-    DeviceStore store = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore store = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     store.addDevice(device);
     assertThat(store.getDevicesList().values(), hasSize(1));
     assertThat(store.getDevicesInUse(), hasSize(0));
@@ -308,7 +319,7 @@ public class DeviceStoreTest {
   @Test
   public void shouldRemoveHardwareDevice() throws Exception {
     DefaultHardwareDevice device = anDevice("de", DeviceTargetPlatform.ANDROID16);
-    DeviceStore store = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore store = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     store.addDevice(device);
     assertThat(store.getDevicesList().values(), hasSize(1));
     store.removeAndroidDevice(device);
@@ -318,7 +329,7 @@ public class DeviceStoreTest {
   @Test
   public void shouldNotRemoveAnEmulator() throws Exception {
     DefaultAndroidEmulator deEmulator10 = anEmulator("de", DeviceTargetPlatform.ANDROID16, false);
-    DeviceStore store = new DeviceStore(false, EMULATOR_PORT);
+    DeviceStore store = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
     store.addEmulators(Arrays.asList(new AndroidEmulator[] {deEmulator10}));
 
     assertThat(store.getDevicesList().values(), hasSize(1));
@@ -328,5 +339,20 @@ public class DeviceStoreTest {
     } catch (DeviceStoreException e) {
       // expected
     }
+  }
+
+  @Test
+  public void shouldFindStartedEmulator() throws Exception {
+    // prepare device store
+    DefaultAndroidEmulator deEmulator16 = anEmulator("de", DeviceTargetPlatform.ANDROID16, true);
+    DeviceStore deviceStore = new DeviceStore(false, EMULATOR_PORT, anDeviceManager());
+    deviceStore.addEmulators(Arrays.asList(new AndroidEmulator[] {deEmulator16}));
+
+    // find by Capabilities
+    AndroidDevice device = deviceStore.findAndroidDevice(withDefaultCapabilities());
+    // The right device is found
+    assertThat(device, equalTo((AndroidDevice) deEmulator16));
+    // the device is in use when found
+    assertThat(deviceStore.getDevicesInUse(), contains((AndroidDevice) deEmulator16));
   }
 }